+++ students/511134962/ood-assignment/src/main/java/com/coderising/ood/srp/common/ConfigurationKeys.java
@@ -1,0 +1,15 @@
+/**********************************************************************************************************************
+ * Copyright (c) 2017. Lorem ipsum dolor sit amet, consectetur adipiscing elit.                                       *
+ * Morbi non lorem porttitor neque feugiat blandit. Ut vitae ipsum eget quam lacinia accumsan.                        *
+ * Etiam sed turpis ac ipsum condimentum fringilla. Maecenas magna.                                                   *
+ * Proin dapibus sapien vel ante. Aliquam erat volutpat. Pellentesque sagittis ligula eget metus.                     *
+ * Vestibulum commodo. Ut rhoncus gravida arcu.                                                                       *
+ **********************************************************************************************************************/
+
+package com.coderising.ood.srp.common;
+
+public class ConfigurationKeys {
+	public static final String SMTP_SERVER = "smtp.server";
+	public static final String ALT_SMTP_SERVER = "alt.smtp.server";
+	public static final String EMAIL_ADMIN = "email.admin";
+}
