+++ students/511134962/ood-assignment/src/main/java/com/coderising/ood/srp/util/DBUtil.java
@@ -6,7 +6,7 @@
  * Vestibulum commodo. Ut rhoncus gravida arcu.                                                                       *
  **********************************************************************************************************************/
 
-package com.coderising.ood.srp;
+package com.coderising.ood.srp.util;
 
 import java.util.ArrayList;
 import java.util.HashMap;
