+++ students/511134962/ood-assignment/src/main/resources/product_promotion.txt
